chore: moves utils to app folder
