--- types/essays.d.ts
@@ -1,7 +1,0 @@
-export type Essay = {
-  date?: string
-  title?: string
-  description?: string
-  urlPath?: string
-  slug?: string
-}
